AIAssistant/backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
google-generativeai>=0.4.0  # <- CORRIGIDO
python-dotenv==1.0.1

